Refactor ApiEngine
